Only automatically override upload settings if a constant is defined
This makes the auto-enabled Azure uploads only trigger when the constant `MICROSOFT_AZURE_FORCE_EXTERNAL_UPLOAD` is both defined, and set to a `truthy` value. By doing this, we enable the highest amount of flexibility to those who wish to use the plugin only for one feature or another (for example only to push Azure settings).